refactoring done, but state not updating
