+++ src/Store.ts
@@ -1,25 +1,6 @@
 import { FeatureCollection } from 'geojson';
+import { ActionHandler, BaseAction, CompositeAction } from './Pipeline';
 
-export interface BaseAction {
-  id: string
-  type: string;
-  label: string;
-  version: string;
-  active: boolean
-}
-
-export interface Action extends BaseAction {
-  payload: any;
-}
-
-export interface CompositeAction extends BaseAction {
-  items: Action[];
-}
-
-export interface ActionHandler {
-  type: string
-  handle(state: FeatureCollection, action: BaseAction): FeatureCollection;
-}
 
 export const TypeComposite = 'composite';
 
@@ -32,22 +13,18 @@ export const printFeature = (msg: string, feature: FeatureCollection) => {
 }
 
 class Store {
-  private actions: Array<Action | CompositeAction>;
   private currentState: FeatureCollection;
   private initialState: FeatureCollection;
   private handlers: ActionHandler[];
   private stateListeners: ((state: FeatureCollection) => void)[];
-  private actionsListeners: ((actions: BaseAction[]) => void)[];
   private index: number
 
   constructor(initialState: FeatureCollection) {
     console.log('store constructor', initialState);
-    this.actions = [];
     this.currentState = initialState;
     this.initialState = initialState;
     this.handlers = []
     this.stateListeners = [];
-    this.actionsListeners = [];
     this.index = 0;
 
     // register the composite handler
@@ -77,15 +54,6 @@ class Store {
     }
   }
 
-  addAction(action: Action | CompositeAction) {
-    const newAction = { ...action };
-    // initialise the id of the action
-    newAction.id = '' + ++this.index
-    this.actions.push(newAction);
-    this.actionsListeners.forEach(listener => listener(this.actions));  
-    this.updateState();
-  }
-
   addHandler(handler: ActionHandler) {
     this.handlers.push(handler);
   }
@@ -94,81 +62,8 @@ class Store {
     this.stateListeners.push(listener)
   }
 
-  addActionsListener(listener: (actions: BaseAction[]) => void) {
-    this.actionsListeners.push(listener)
-  }
-
-  removeAction(action: BaseAction) {
-    // check it's in the top level actions
-    if (this.actions.includes(action as CompositeAction | Action)) {
-      this.actions = this.actions.filter(a => a !== action);
-    } else {
-      // check inside composite actions
-      const compositeActions = this.actions.filter(a => a.type === TypeComposite) as CompositeAction[];
-      const parent = compositeActions.find(comp => comp.items.includes(action as Action));
-      if (parent) {
-        parent.items = parent?.items.filter(a => a !== action);
-      }
-    }
-
-    this.actions = this.actions.filter(a => a !== action);
-    this.actionsListeners.forEach(listener => listener(this.actions));
-    this.updateState();
-  }
-
-  ungroupAction(action: BaseAction) {
-    // check it's a composite action
-    if (action.type !== TypeComposite) {
-      console.warn('Action is not a composite action', action);
-      return;
-    }
-    else {
-      const comp = action as CompositeAction; 
-      // find the index of the composite action
-      const index = this.actions.findIndex(a => a === action);
-      // remove the composite action
-      this.actions.splice(index, 1);
-      // add the items back to the action list at the correct index
-      this.actions.splice(index, 0, ...comp.items);
-      // inform the action listeners
-      this.actionsListeners.forEach(listener => listener(this.actions));
-      // update the state
-      this.updateState();
-    }  
-  }
-
-  groupActions(selectedItems: BaseAction[], name: string) {
-    const compositeAction: CompositeAction = {
-      id: '' + ++this.index,
-      type: TypeComposite,
-      label: name,
-      version: '1.0',
-      active: true,
-      items: selectedItems as Action[]
-    }
-    // find the index of the first action in the selected items
-    const firstIndex = this.actions.findIndex(action => action === selectedItems[0]);
-
-    // insert the composite action before the first action
-    this.actions.splice(firstIndex, 0, compositeAction);
-
-    // filter out the selected items
-    this.actions = this.actions.filter(action => !selectedItems.includes(action));
-
-    // fire updates to action list listeners
-    this.actionsListeners.forEach(listener => listener(this.actions));
-
-    // update the state
-    this.updateState();
-  }
-
-  toggleActionActive(action: BaseAction) {
-    action.active = !action.active;
-    this.updateState();
-  }
-
-  private updateState() {
-    this.currentState = this.actions.reduce((state, action) => {
+  private updateState(actions: BaseAction[]) {
+    this.currentState = actions.reduce((state, action) => {
       if (!action.active) {
         return state;
       }
